refactor(engine-api): fold phosphor-engine-types headers into phosphor-engine-api
Phase A2 of docs/phosphor-architecture-refactor-plan.md.

The phosphor-engine-types lib was INTERFACE-only and its namespace was
already PhosphorEngineApi — the split was an artifact and the
namespace/lib mismatch was going to be a rebrand trap.

Implementation note: phosphor-engine-api had stub forwarder headers
(IPlacementState.h, NavigationContext.h, EngineTypes.h) that just
re-included the real content from <PhosphorEngineTypes/...>. Rather
than coexisting two parallel headers, this collapses the stub
forwarders by inlining the real content into the engine-api headers
under namespace PhosphorEngineApi. Single source of truth, no stale
forwarders.

Files removed:
- libs/phosphor-engine-types/ (entire lib)

Files updated (include rewrites <PhosphorEngineTypes/...> →
<PhosphorEngineApi/...>):
- libs/phosphor-snap-engine/include/PhosphorSnapEngine/{ISnapSettings,SnapEngine,SnapState}.h
- libs/phosphor-snap-engine/src/navigation.cpp
- libs/phosphor-tile-engine/include/PhosphorTileEngine/{AutotileEngine,IAutotileSettings}.h
- libs/phosphor-tiles/include/PhosphorTiles/TilingState.h
- src/core/{constants,enums}.h

CMake updates:
- libs/phosphor-engine-api/CMakeLists.txt — drops the
  PhosphorEngineTypes link dep
- libs/phosphor-snap-engine/{CMakeLists.txt,Config.cmake.in} — same
- libs/phosphor-tile-engine/Config.cmake.in — same
- libs/phosphor-tiles/{CMakeLists.txt,Config.cmake.in} — same
- libs/CMakeLists.txt — drops add_subdirectory(phosphor-engine-types)

163/163 tests pass.
