+++ libs/phosphor-engine-api/CMakeLists.txt
@@ -53,7 +53,6 @@ target_link_libraries(PhosphorEngineApi
     PUBLIC
         Qt6::Core
         Qt6::Gui
-        PhosphorEngineTypes::PhosphorEngineTypes
         PhosphorGeometry::PhosphorGeometry
 )
