+++ libs/phosphor-engine-api/include/PhosphorEngineApi/NavigationContext.h
@@ -3,4 +3,21 @@
 
 #pragma once
 
-#include <PhosphorEngineTypes/NavigationContext.h>
+#include <QString>
+
+namespace PhosphorEngineApi {
+
+/// Target window + screen for a navigation or lifecycle operation.
+///
+/// Populated by the daemon from compositor shadow state (windowActivated /
+/// cursorScreenChanged) before dispatching through the placement engine.
+/// Both fields may be empty on very-early-startup shortcuts or when no
+/// window is focused — each IPlacementEngine method documents its
+/// behaviour when the fields are empty.
+struct NavigationContext
+{
+    QString windowId;
+    QString screenId;
+};
+
+} // namespace PhosphorEngineApi
