+++ libs/phosphor-snap-engine/include/PhosphorSnapEngine/ISnapSettings.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include <PhosphorEngineTypes/EngineTypes.h>
+#include <PhosphorEngineApi/EngineTypes.h>
 #include <QObject>
 #include <QStringList>
