+++ libs/phosphor-tiles/PhosphorTilesConfig.cmake.in
@@ -6,6 +6,7 @@
 include(CMakeFindDependencyMacro)
 
 find_dependency(Qt6 6.6 COMPONENTS Core Qml)
+find_dependency(PhosphorEngineApi)
 find_dependency(PhosphorLayoutApi)
 
 include("${CMAKE_CURRENT_LIST_DIR}/PhosphorTilesTargets.cmake")
